Change internal dependencies from requirejs to DI.
Stubbing requirejs modules is not an easy task.  This is due in part to
the asynchronous nature of requirejs and in part to the difficulty in
defining a different module for stubbing other tests.  The easiest way
to take care of this would seem to be to just structure the app so that
dependencies have to be injected for each module.