Use mongo id instead of title for identifying posts.
Using a user-created field as the identifier in Backbone causes
headaches.  Even though it looks worse, it's better from a development
standpoint to use the mongo identifier.  A possible solution would be to
change the api to create the mongo id as a url friendly version of the
title.  This way backbone would be oblivious to the id, but we still get
a decent looking url.